perf: serve public app shell statically
