+++ src/app/layout.tsx
@@ -1,13 +1,8 @@
 import type { Metadata, Viewport } from "next";
-import { cookies } from "next/headers";
 import "./globals.css";
-import { NextIntlClientProvider } from "next-intl";
-import { getLocale } from "next-intl/server";
 import ThemeBootstrap from "@/components/ThemeBootstrap";
 import { Toaster } from "@/components/ui/sonner";
 import { TooltipProvider } from "@/components/ui/tooltip";
-import { isValidLocale } from "@/lib/i18n/locales";
-import { pickCatalogNamespaces } from "@/i18n/catalogs";
 import {
   DEFAULT_SOCIAL_IMAGE,
   DEFAULT_SOCIAL_IMAGE_HEIGHT,
@@ -20,14 +15,7 @@ import {
   SITE_TITLE,
   getSiteUrl,
 } from "@/lib/seo";
-import {
-  parseResolvedTheme,
-  parseThemePreference,
-  RESOLVED_THEME_COOKIE,
-  resolveTheme,
-  THEME_COOKIE,
-  type ResolvedTheme,
-} from "@/lib/theme";
+import { THEME_BOOTSTRAP_SCRIPT } from "@/lib/theme";
 
 export const metadata: Metadata = {
   metadataBase: new URL(getSiteUrl()),
@@ -74,51 +62,24 @@ export const viewport: Viewport = {
   themeColor: "#0c0c0f",
 };
 
-function getInitialTheme(
-  preferenceCookie: string | undefined,
-  resolvedCookie: string | undefined
-): ResolvedTheme {
-  const preference = parseThemePreference(preferenceCookie);
-  const resolved = parseResolvedTheme(resolvedCookie);
-
-  if (preference === "light" || preference === "dark") {
-    return preference;
-  }
-
-  if (resolved) {
-    return resolved;
-  }
-
-  return resolveTheme("system", false);
-}
-
-export default async function RootLayout({
+export default function RootLayout({
   children,
 }: {
   children: React.ReactNode;
 }) {
-  const cookieStore = await cookies();
-  const initialTheme = getInitialTheme(
-    cookieStore.get(THEME_COOKIE)?.value,
-    cookieStore.get(RESOLVED_THEME_COOKIE)?.value
-  );
-  const rawLocale = await getLocale();
-  const locale = isValidLocale(rawLocale) ? rawLocale : "en";
-  const messages = await pickCatalogNamespaces(locale, ["common"]);
-
   return (
     <html
-      lang={locale}
-      className={initialTheme === "dark" ? "dark" : undefined}
-      style={{ colorScheme: initialTheme }}
+      lang="en"
+      style={{ colorScheme: "light" }}
       data-scroll-behavior="smooth"
       suppressHydrationWarning
     >
+      <head>
+        <script dangerouslySetInnerHTML={{ __html: THEME_BOOTSTRAP_SCRIPT }} />
+      </head>
       <body className="font-sans antialiased">
         <ThemeBootstrap />
-        <NextIntlClientProvider locale={locale} messages={messages}>
-          <TooltipProvider delayDuration={500}>{children}</TooltipProvider>
-        </NextIntlClientProvider>
+        <TooltipProvider delayDuration={500}>{children}</TooltipProvider>
         <Toaster position="bottom-right" richColors />
       </body>
     </html>
