+++ src/app/terms/page.tsx
@@ -1,23 +1,12 @@
-import type { Metadata } from "next";
+"use client";
+
 import Link from "next/link";
-import { getTranslations } from "next-intl/server";
+import { useTranslations } from "next-intl";
 import { Footer } from "@/components/landing/Footer";
 import { PublicSiteHeader } from "@/components/landing/PublicSiteHeader";
-import LanguageProvider from "@/i18n/LanguageProvider";
 
-export async function generateMetadata(): Promise<Metadata> {
-  const t = await getTranslations("legal");
-  return {
-    title: t("terms.metaTitle"),
-    description: t("terms.metaDescription"),
-    alternates: {
-      canonical: "/terms",
-    },
-  };
-}
-
-export default async function TermsPage() {
-  const t = await getTranslations("legal");
+export default function TermsPage() {
+  const t = useTranslations("legal");
 
   const sections = [
     {
@@ -59,69 +48,67 @@ export default async function TermsPage() {
   ];
 
   return (
-    <LanguageProvider namespaces={["common", "landing"]}>
-      <div className="bg-background text-foreground min-h-screen">
-        <PublicSiteHeader currentPage="legal" />
+    <div className="bg-background text-foreground min-h-screen">
+      <PublicSiteHeader currentPage="legal" />
 
-        <main>
-          <section className="border-border/40 border-b">
-            <div className="mx-auto w-full max-w-4xl px-6 py-14 sm:py-20">
-              <p className="text-brand-primary text-[11px] font-semibold tracking-[0.2em] uppercase">
-                {t("terms.breadcrumbLegal")}
-              </p>
-              <h1 className="mt-4 text-4xl leading-tight font-semibold tracking-normal sm:text-5xl">
-                {t("terms.pageTitle")}
-              </h1>
-              <p className="text-muted-foreground mt-5 max-w-2xl text-sm leading-7">
-                {t("terms.effectiveDate")}
-              </p>
-            </div>
-          </section>
+      <main>
+        <section className="border-border/40 border-b">
+          <div className="mx-auto w-full max-w-4xl px-6 py-14 sm:py-20">
+            <p className="text-brand-primary text-[11px] font-semibold tracking-[0.2em] uppercase">
+              {t("terms.breadcrumbLegal")}
+            </p>
+            <h1 className="mt-4 text-4xl leading-tight font-semibold tracking-normal sm:text-5xl">
+              {t("terms.pageTitle")}
+            </h1>
+            <p className="text-muted-foreground mt-5 max-w-2xl text-sm leading-7">
+              {t("terms.effectiveDate")}
+            </p>
+          </div>
+        </section>
 
-          <section>
-            <div className="mx-auto w-full max-w-4xl px-6 py-12 sm:py-16">
-              <div className="space-y-10">
-                {sections.map((section) => (
-                  <section key={section.title}>
-                    <h2 className="text-xl font-semibold tracking-normal">
-                      {section.title}
-                    </h2>
-                    <div className="text-muted-foreground mt-3 space-y-3 text-sm leading-7">
-                      {section.body.map((paragraph, i) => (
-                        <p key={i}>{paragraph}</p>
-                      ))}
-                    </div>
-                  </section>
-                ))}
-
-                <section className="border-border/50 bg-muted/18 rounded-2xl border p-6">
+        <section>
+          <div className="mx-auto w-full max-w-4xl px-6 py-12 sm:py-16">
+            <div className="space-y-10">
+              {sections.map((section) => (
+                <section key={section.title}>
                   <h2 className="text-xl font-semibold tracking-normal">
-                    {t("terms.s10Title")}
+                    {section.title}
                   </h2>
-                  <p className="text-muted-foreground mt-3 text-sm leading-7">
-                    {t("terms.s10Body")}
-                  </p>
+                  <div className="text-muted-foreground mt-3 space-y-3 text-sm leading-7">
+                    {section.body.map((paragraph, i) => (
+                      <p key={i}>{paragraph}</p>
+                    ))}
+                  </div>
                 </section>
+              ))}
 
-                <p className="text-muted-foreground text-sm leading-7">
-                  {t.rich("terms.seePrivacy", {
-                    link: (chunks) => (
-                      <Link
-                        href="/privacy"
-                        className="text-brand-primary hover:text-brand-primary/85 font-medium transition-colors"
-                      >
-                        {chunks}
-                      </Link>
-                    ),
-                  })}
+              <section className="border-border/50 bg-muted/18 rounded-2xl border p-6">
+                <h2 className="text-xl font-semibold tracking-normal">
+                  {t("terms.s10Title")}
+                </h2>
+                <p className="text-muted-foreground mt-3 text-sm leading-7">
+                  {t("terms.s10Body")}
                 </p>
-              </div>
-            </div>
-          </section>
-        </main>
+              </section>
 
-        <Footer />
-      </div>
-    </LanguageProvider>
+              <p className="text-muted-foreground text-sm leading-7">
+                {t.rich("terms.seePrivacy", {
+                  link: (chunks) => (
+                    <Link
+                      href="/privacy"
+                      className="text-brand-primary hover:text-brand-primary/85 font-medium transition-colors"
+                    >
+                      {chunks}
+                    </Link>
+                  ),
+                })}
+              </p>
+            </div>
+          </div>
+        </section>
+      </main>
+
+      <Footer />
+    </div>
   );
 }
