+++ wrangler.jsonc
@@ -20,6 +20,7 @@
   "assets": {
     "directory": ".open-next/assets",
     "binding": "ASSETS",
+    "run_worker_first": false,
   },
   "r2_buckets": [
     {
